pyproject.toml
[project]
name = "ageoke-v2"
version = "2.0.0"
description = "Adaptive Geological Knowledge Extraction - Layout-aware PDF extraction pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Observability / tracing
    "arize-phoenix>=13.1.1",
    "openinference-instrumentation-langchain>=0.1.59",
    "openinference-instrumentation-openai>=0.1.41",
    # LLM orchestration (workflow / agent layer)
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    # Structured LLM output (extraction layer)
    "instructor>=1.7.0",
    "openai>=1.0.0",
    # Layout-aware PDF parsing (replaces Tesseract OCR)
    "docling>=2.0.0",
    # Data validation
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    # MCP server infrastructure
    "fastmcp>=2.0.0",
    # Mindat database API
    "openmindat>=0.1.3",
    # CLI
    "typer>=0.12.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]
